Move cacheable_label_details from labels/models.py to utils.py, and have it count confirmed annotations, not all annotations.
