+++ project/labels/tests/test_label_main.py
@@ -20,6 +20,7 @@ from visualization.utils import get_patch_path
 from ..models import LabelGroup, Label
 from ..templatetags.labels import (
     popularity_bar as popularity_bar_tag, status_icon as status_icon_tag)
+from ..utils import label_popularity
 
 
 class PermissionTest(BasePermissionTest):
@@ -233,7 +234,7 @@ class LabelMainTest(BaseLabelMainTest):
 
         # Popularity.
         with context_scoped_cache():
-            popularity_str = str(int(label_a.popularity)) + '%'
+            popularity_str = str(int(label_popularity(label_a.pk))) + '%'
             popularity_bar_html = popularity_bar_tag(label_a)
         self.assertInHTML(
             'Popularity: {} {}'.format(
@@ -549,7 +550,7 @@ class PopularityTest(ClientTest):
 
         with context_scoped_cache():
             self.assertEqual(
-                self.label_a.popularity, 0,
+                label_popularity(self.label_a.pk), 0,
                 msg="0 sources should mean 0 popularity")
 
     def test_zero_annotations(self):
@@ -561,7 +562,7 @@ class PopularityTest(ClientTest):
 
         with context_scoped_cache():
             self.assertEqual(
-                self.label_a.popularity, 0,
+                label_popularity(self.label_a.pk), 0,
                 msg="1 source and 0 annotations still should mean 0 popularity")
 
     def test_nonzero_annotations(self):
@@ -574,7 +575,7 @@ class PopularityTest(ClientTest):
 
         with context_scoped_cache():
             self.assertGreater(
-                self.label_a.popularity, 0,
+                label_popularity(self.label_a.pk), 0,
                 msg="Non-0 annotations should mean non-0 popularity")
